Port the 'finding your way' docs
This is a direct port of the decision tree that @jtigger implemented to help people
orient themselves when they first start contributing to Exercism.

I put the very general 'help us make this better' stuff in the top-level README. I also
swapped out the images with octicons, to make it more visually consistent.

We may want to consider breaking this out slightly differently once we have gotten all
the documentation into one place.