Add documents resulting from first round of user experience design
These are philosophical in nature.

Once we have settled on what we believe is the correct foundation, we
will move forward and explore the practical implications of this.

We are trying to preserve the core goodness that Exercism has, while
addressing the areas where we have been less successful.

In some areas there will not be very large changes, whereas in others we
anticipate having to do quite a lot of work to support the goals of the
project, the participants, and their success.