exercises: Document `title`
In generated READMEs, the rule for converting an exercise slug to a
human-readable form is to capitalise only the first letter of each
hyphen-separated word.

For some exercises this is not the best possible title. For example see
http://exercism.io/exercises/haskell/rna-transcription/readme and
observe that the title is "Rna Transcription", whereas "RNA
Transcription" seems more faithful to the rules (RNA, as an initialism,
should be in all caps).

For those exercises, we add an optional `title` field in metadata.yml to
allow overriding the default capitalisation or punctuation.

This documents this field and how it is used in README generation.

Discussion:
https://github.com/exercism/configlet/issues/76

Addition of `title` field:
https://github.com/exercism/problem-specifications/pull/901

Implementation in `configlet`:
https://github.com/exercism/configlet/pull/85