Update "implement an exercise from specification"
The "implement an exercise from specification" documentation was
slightly out of date in its description of the exercises object.
This has been updated with a link to the exercises.md document which
explains in detail what should be in the exercises object (in
config.json).

Also, the exercises.md document didn't have a link to the topics list
so that has been added.