Update code review philosophy doc
We don't need a deep knowledge of the language, just an adequate one.
It's particularly important if someone has a weaker understanding of
the language, that they are open, honest, humble about it, and that
they approach the conversations with curiosity, and that they're
comfortable reaching out to the other mentors for help and feedback.